iio_info: Remove -s option
iio_info had two possible options to scan for contexts: -S (aka. --scan)
and -s, which did more or less the same thing, but -s didn't take an
argument and scanned all possible contexts.

Remove the -s option in the help text to reduce the confusion, while
still handling it for backwards-compatibility.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>