utils: iio_rwdev: try clean exit before cancelling the buffer
When destroying the stream, we'll try to dequeue (as it improves things
on iiod side) any possible pending block and if the buffer is already
cancelled we'll just timeout and make  exiting the app much slower. Hence,
let's try a clean exit first and if we get blocked a second hit on ctrl-c
will go and cancel the buffer as before.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>