zephyr: Add usb cdc-acm snippet
Adds a new zephyr build system snippet that enables the iiod server over
a USB CDC ACM UART. This makes it easy to add iiod to an existing
application without modifying the application. For example, the
following build commands demonstrate using the `-S iiod-cdc-acm`
argument to extend three different in-tree sample applications on three
different boards:

$ west build -b apard32690/max32690/m4 \
	~/zephyrproject/zephyr/samples/synchronization/ \
	-S iiod-cdc-acm

$ west build -b frdm_k64f \
	~/zephyrproject/zephyr/samples/subsys/shell/shell_module/ \
	-S iiod-cdc-acm

$ west build -b adi_sdp_k1 \
	~/zephyrproject/zephyr/samples/basic/blinky \
	-S iiod-cdc-acm

The console and iiod are directed to separate uart devices, therefore
it's possible to communicate with both simultaneously.

Signed-off-by: Maureen Helm <maureen.helm@analog.com>